[ADD] support crop by original ratio

+++ FMPhotoPicker/FMPhotoPicker/source/ImageEditor/Crop/FMCropName.swift
@@ -9,8 +9,8 @@
 import UIKit
 
 struct FMCropRatio {
-    let width: Int
-    let height: Int
+    let width: CGFloat
+    let height: CGFloat
 }
 
 enum FMCropName: String {
